+++ testsuite/integration-arquillian/tests/base/src/main/java/org/keycloak/testsuite/arquillian/containers/KeycloakContainerTestExtension.java
@@ -49,7 +49,6 @@ import org.jboss.arquillian.test.impl.context.TestContextImpl;
 import org.jboss.arquillian.test.impl.enricher.resource.ArquillianResourceTestEnricher;
 import org.jboss.arquillian.test.spi.TestEnricher;
 import org.jboss.arquillian.test.spi.enricher.resource.ResourceProvider;
-import org.keycloak.testsuite.arquillian.ModelTestExecutor;
 
 /**
  * KeycloakContainerTestExtension
@@ -105,8 +104,7 @@ public class KeycloakContainerTestExtension implements LoadableExtension {
             .observer(ClientDeployerCreator.class)
             .observer(ClientBeforeAfterLifecycleEventExecuter.class)
             .observer(ClientTestExecuter.class)
-//            .observer(LocalTestExecuter.class)
-            .observer(ModelTestExecutor.class)
+            .observer(LocalTestExecuter.class)
             .observer(RemoteTestExecuter.class)
             .observer(DeploymentCommandObserver.class)
             .observer(ContainerCommandObserver.class)
