+++ testsuite/integration-arquillian/tests/base/src/main/java/org/keycloak/testsuite/client/resources/TestingResource.java
@@ -290,13 +290,6 @@ public interface TestingResource {
     @Produces(MediaType.TEXT_PLAIN_UTF_8)
     String runOnServer(String runOnServer);
 
-    @POST
-    @Path("/run-model-test-on-server")
-    @Consumes(MediaType.TEXT_PLAIN_UTF_8)
-    @Produces(MediaType.TEXT_PLAIN_UTF_8)
-    String runModelTestOnServer(@QueryParam("testClassName") String testClassName,
-                                @QueryParam("testMethodName") String testMethodName);
-
     @GET
     @Path("js/keycloak.js")
     @Produces(MediaType.TEXT_HTML_UTF_8)
