+++ testsuite/integration-deprecated/src/test/java/org/keycloak/testsuite/model/ClientModelTest.java
@@ -6,7 +6,6 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- * 
  * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
